+++ yash-env/CHANGELOG.md
@@ -20,6 +20,9 @@ A _private dependency_ is used internally and not visible to downstream users.
   in the real system implementation.
 - `impl system::Stat for system::real::Stat`
 - `impl system::Stat for system::virtual::Stat`
+- The `awaiters: Vec<Waker>` field has been added to the
+  `system::virtual::FileBody::Fifo` variant to allow the virtual system to
+  notify tasks waiting on the pipe when certain events occur.
 
 ### Changed